# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample QRaven API.",
        "title": "QRaven API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/ticket/{eventId}": {
            "post": {
                "description": "Creates a new ticket for the event with the given eventId.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "Creates a new ticket for an event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create Ticket Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateTicketRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ticketResponse",
                        "schema": {
                            "$ref": "#/definitions/models.CreateTicketResponse"
                        }
                    },
                    "400": {
                        "description": "badRequest",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "422": {
                        "description": "validationError",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CreateTicketRequest": {
            "type": "object",
            "required": [
                "amount",
                "event_id",
                "type",
                "user_id"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "event_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.CreateTicketResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "event_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "purchase_time": {
                    "type": "string"
                },
                "qr_code": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "description": "for errors that occur even if request is successful"
                },
                "extra": {},
                "message": {
                    "type": "string"
                },
                "name": {
                    "description": "name of the error",
                    "type": "string"
                },
                "pagination": {},
                "status": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        }
    }
}